Re-Release 1.1.0, Fix param order & documentation
Upon updating the documentation, I discovered that the variable order
for `delegateWithParamMask` was "unnatural" compared to how functions
are called in PHP.  Therefore I deprecated 1.1.0 and am re-releasing it
as 1.1.1.  Due to the short time frame (within an hour), it is unlikely
anyone used 1.1.0 and so would be affected by the breaking change.